Move prepareCookieStats, prepareFrameStats and prepareCookiesCount to design-system.
