+++ packages/design-system/src/components/circle/stories/circle.stories.tsx
@@ -17,7 +17,10 @@
  * External dependencies.
  */
 import type { Meta, StoryObj } from '@storybook/react';
-import { COLOR_MAP } from '@ps-analysis-tool/common';
+/**
+ * Internal dependencies
+ */
+import { COLOR_MAP } from '../../../constants';
 
 /**
  * Internal dependencies.
