+++ packages/design-system/src/components/matrix/matrixComponent/stories/MatrixComponent.stories.tsx
@@ -17,12 +17,12 @@
  * External dependencies.
  */
 import type { Meta, StoryObj } from '@storybook/react';
-import { COLOR_MAP } from '@ps-analysis-tool/common';
 
 /**
  * Internal dependencies.
  */
 import MatrixComponent from '..';
+import { COLOR_MAP } from '../../../../constants';
 
 const meta: Meta<typeof MatrixComponent> = {
   title: 'DesignSystem/Matrix/Component',
