+++ packages/design-system/src/index.ts
@@ -17,3 +17,4 @@
 export * from './icons';
 export * from './components';
 export * from './utils';
+export * from './constants';
