+++ packages/design-system/src/utils/index.ts
@@ -15,3 +15,6 @@
  */
 
 export { noop } from './noop';
+export { default as prepareCookiesCount } from './prepareCookiesCount';
+export { default as prepareCookieStatsComponents } from './prepareCookieStatsComponents';
+export { default as prepareFrameStatsComponent } from './prepareFrameStatsComponent';
